fixes for some corner cases from Ilja van Sprundel:     - tempnam     - strsignal     - strftime     - sunrpc code

+++ libugly/strsignal.c
@@ -3,7 +3,7 @@
 #include <signal.h>
 
 const char* strsignal(int sig) {
-  if (sig<=SIGRTMAX)
+  if ((unsigned int)sig<=SIGRTMAX)
     return sys_siglist[sig];
   else
     return "(unknown signal)";
